+++ functions/response_test.ts
@@ -1,7 +1,7 @@
 import { SlackFunctionTester } from "deno-slack-sdk/mod.ts";
 import { assertEquals } from "std/testing/asserts.ts";
 import ResponseFunction from "./response.ts";
-import env from "../env.ts";
+import conf from "../conf.ts";
 
 const { createContext } = SlackFunctionTester("response");
 
@@ -10,7 +10,7 @@ const { createContext } = SlackFunctionTester("response");
 Deno.test("Response function test -- keyword", async () => {
   const inputs = { message: `<ABCDEFGHIJK> 今日花金？` };
   const { outputs } = await ResponseFunction(createContext({ inputs }));
-  assertEquals(outputs?.response, env.usage);
+  assertEquals(outputs?.response, conf.usage);
 });
 
 // when: @hanakin keyword on Friday
@@ -103,7 +103,7 @@ Deno.test("Response function test -- non-keyword", async () => {
 
   const inputs = { message: `<@ABCDEFGHIJK> ABCDEFGH` };
   const { outputs } = await ResponseFunction(createContext({ inputs }));
-  assertEquals(outputs?.response, env.usage);
+  assertEquals(outputs?.response, conf.usage);
 });
 
 // when: @hanakin 今日目黒で花金?
@@ -113,7 +113,7 @@ Deno.test("Response function test -- keyword with place", async () => {
 
   // generate test data
   const answerArray: string[] = [];
-  const array = env.recommended_bar["目黒"];
+  const array = conf.recommended_bar["目黒"];
   array.forEach((e) => {
     answerArray.push(
       `今日は花金！${e.name}で${e.main}を飲もう！${e.url}`,
@@ -135,6 +135,6 @@ Deno.test("Response function test -- keyword with wrong place", async () => {
   const { outputs } = await ResponseFunction(createContext({ inputs }));
   assertEquals(
     outputs?.response,
-    "福岡は登録されていないみたいよ。https://github.com/chaspy/hanakintaro/blob/main/env.ts におすすめの店を追加しよう",
+    "福岡は登録されていないみたいよ。https://github.com/chaspy/hanakintaro/blob/main/conf.ts におすすめの店を追加しよう",
   );
 });
